Created components as sub components of Account.js
Created UpdateEmail.js and UpdatePword.js, added open/close button for each component.
